Feat: 0829 Cart 구현 및 Address 고침

+++ mini/src/main/java/com/example/mini/entity/Product.java
@@ -22,13 +22,13 @@ import lombok.Setter;
 public class Product {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	
 	@Column(length = 50)
 	private String product_name;
 	@Column(length = 300)
 	private String description;
-	private Integer product_price;
+	private Long product_price;
 	private String product_quantity;
 	private String image_url; 
 	@OneToMany(cascade = CascadeType.REMOVE)
